wip
FossilOrigin-Name: 771f3ff9cdd7f711ce50c4b879518cbac485570602dc3522033dc5d1af2e1aaf

+++ include/cwutils/cwarray.h
@@ -12,17 +12,13 @@ typedef struct CwArray {
 } CwArray;
 
 CwArray* cwarray_new(size_t element_size);
-void* cwarray_push(CwArray* self);
-void* cwarray_get(CwArray* self, size_t index);
-void* cwarray_demote(CwArray* self);
+void*    cwarray_push(CwArray* self);
+void*    cwarray_get(CwArray* self, size_t index);
+void*    cwarray_demote(CwArray* self);
 
 int cwarray_size(CwArray* self);
-
 CwArray* cwarray_free(CwArray* self);
-
 void cwarray_for_each(CwArray* self, void (*func)(void*));
-
-
 
 // A CwList is a special case of a CwArray where element_size
 // is sizeof(void*).
@@ -32,6 +28,6 @@ CwList* cwlist_new();
 CwList* cwlist_with_elements(size_t size, ...);
 void cwlist_push(CwList* self, void* item);
 void cwlist_push_many(CwList* self, size_t size, ...);
-void* cwlist_get(CwList* self, size_t index);
+void* cwlist_get(CwList* self, int index);
 
 #endif
